+++ Model/FreightRide.cs
@@ -2,22 +2,25 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace T_HUB.Model {
-    class FreightRide : Ride {
+namespace T_HUB.Model
+{
+    class FreightRide : Ride
+    {
 
         /// <summary>
         /// Total weigth of the goods
         /// </summary>
-        public float Wg { get; }
+        public double Wg { get; }
 
         /// <summary>
         /// Volume occupied by the goods
         /// </summary>
-        public float Vol { get; }
+        public double Vol { get; }
 
-        public FreightRide(string vehType, string vehLicPlt, float km, DateTime startTm, 
-            DateTime endTm, float wg, float vol, float startPrc = 0) : 
-            base(vehType, vehLicPlt, km, startTm, endTm, startPrc) {
+        public FreightRide(string vehType, string vehLicPlt, double km, DateTime startTm,
+            DateTime endTm, double wg, double vol, double startPrc = 0) :
+            base(vehType, vehLicPlt, km, startTm, endTm, startPrc)
+        {
             this.Wg = wg;
             this.Vol = vol;
         }
